requirements.txt
flask==3.0.0
msal==1.26.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.6.1
anthropic==0.8.1
google-generativeai
boto3
